Rish (#5)
Version 2 with secure folder verification.

The only usecase that I couldn't fix was when using the patch "chang packag name" and installing two apps from the same repo but with different package names. It's not something I wanna think about, and the logs are clear if someone tries to do it.